Refactor #169 - rename `Icons` folder
