+++ app/src/main/java/com/nlab/reminder/domain/feature/home/icon/IcHomeCategoryTimetable.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.domain.feature.home.icons
+package com.nlab.reminder.domain.feature.home.icon
 
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.PathFillType.Companion.NonZero
@@ -25,7 +25,7 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.graphics.vector.ImageVector.Builder
 import androidx.compose.ui.graphics.vector.path
 import androidx.compose.ui.unit.dp
-import com.nlab.reminder.core.android.designsystem.component.ReminderIcons
+import com.nlab.reminder.core.android.designsystem.icon.ReminderIcons
 
 /**
  * @author Doohyun
